Exercise 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,5 @@
 const blogRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
-const jwt = require('jsonwebtoken');
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -10,13 +8,7 @@ blogRouter.get('/', async (request, response) => {
 
 blogRouter.post('/', async (request, response, next) => {
   const body = request.body;
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token is invalid' });
-  }
-
-  const user = await User.findById(request.decodedToken.id);
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
@@ -71,8 +63,20 @@ blogRouter.put('/:id', (request, response, next) => {
 });
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const id = request.params.id;
+  if (id.match(/^[0-9a-fA-F]{24}$/)) {
+    const user = request.user;
+    const blog = await Blog.findById(id);
+
+    if (user && blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      response.status(401).send({ error: 'Unauthorized: invalid token' });
+    }
+  } else {
+    response.status(400).end();
+  }
 });
 
 module.exports = blogRouter;
